+++ src/client/app/services/aws-cognito.service.ts
@@ -7,7 +7,7 @@ import * as moment from 'moment';
 import { environment } from '../../environments/environment';
 
 @Injectable({
-    providedIn: 'root'
+    providedIn: 'root',
 })
 export class AwsCognitoService {
     public static REGION: string = environment.REGION;
@@ -28,23 +28,204 @@ export class AwsCognitoService {
         if (this.isAuthenticate()) {
             return;
         }
-        AWS.config.region = AwsCognitoService.REGION;
+        AWS.config.update({ region: AwsCognitoService.REGION });
         const deviceId = localStorage.getItem('deviceId');
         if (deviceId === 'undefined' || deviceId === null) {
             AWS.config.credentials = new AWS.CognitoIdentityCredentials({
-                IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID
+                IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID,
             });
         } else {
             AWS.config.credentials = new AWS.CognitoIdentityCredentials({
                 IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID,
-                IdentityId: deviceId
+                IdentityId: deviceId,
             });
         }
-        this.credentials = (<AWS.CognitoIdentityCredentials>AWS.config.credentials);
+        this.credentials = <AWS.CognitoIdentityCredentials>(
+            AWS.config.credentials
+        );
         await this.credentials.getPromise();
         if (deviceId === 'undefined' || deviceId === null) {
             localStorage.setItem('deviceId', this.credentials.identityId);
         }
+    }
+
+    /**
+     * ユーザー取得
+     */
+    private async getUser(params: {
+        accessToken: string;
+    }): Promise<AWS.CognitoIdentityServiceProvider.GetUserResponse> {
+        return new Promise((resolve, reject) => {
+            AWS.config.update({ region: AwsCognitoService.REGION });
+            const { accessToken } = params;
+            const cognitoIdentityServiceProvider =
+                new AWS.CognitoIdentityServiceProvider();
+            cognitoIdentityServiceProvider.getUser(
+                { AccessToken: accessToken },
+                (err, data) => {
+                    if (err) {
+                        reject(err);
+                        return;
+                    }
+                    resolve(data);
+                }
+            );
+        });
+    }
+
+    /**
+     * プロフィール取得
+     */
+    public async getProfile(params: { accessToken: string }): Promise<{
+        additionalProperty?: { name: string; value: string }[];
+        email?: string;
+        givenName?: string;
+        familyName?: string;
+        telephone?: string;
+    }> {
+        const { accessToken } = params;
+        const user = await this.getUser({ accessToken });
+        const profile: {
+            additionalProperty?: { name: string; value: string }[];
+            email?: string;
+            givenName?: string;
+            familyName?: string;
+            telephone?: string;
+        } = {
+            additionalProperty: undefined,
+            email: undefined,
+            givenName: undefined,
+            familyName: undefined,
+            telephone: undefined,
+        };
+        profile.additionalProperty = user.UserAttributes.filter(
+            (a) => a.Value !== undefined
+        ).map((a) => ({
+            name: a.Name,
+            value: <string>a.Value,
+        }));
+        user.UserAttributes.forEach((a) => {
+            if (a.Name === 'email') {
+                profile[a.Name] = a.Value;
+            }
+            if (a.Name === 'phone_number') {
+                profile.telephone = a.Value;
+            }
+            if (a.Name === 'family_name') {
+                profile.familyName = a.Value;
+            }
+            if (a.Name === 'given_name') {
+                profile.givenName = a.Value;
+            }
+        });
+        return profile;
+    }
+
+    /**
+     * 属性更新
+     */
+    private async updateUserAttributes(params: {
+        accessToken: string;
+        userAttributes: AWS.CognitoIdentityServiceProvider.AttributeListType;
+    }): Promise<AWS.CognitoIdentityServiceProvider.UpdateUserAttributesResponse> {
+        return new Promise((resolve, reject) => {
+            AWS.config.update({ region: AwsCognitoService.REGION });
+            const { accessToken, userAttributes } = params;
+            const cognitoIdentityServiceProvider =
+                new AWS.CognitoIdentityServiceProvider();
+            cognitoIdentityServiceProvider.updateUserAttributes(
+                {
+                    AccessToken: accessToken,
+                    UserAttributes: userAttributes,
+                },
+                (err, data) => {
+                    if (err) {
+                        reject(err);
+                        return;
+                    }
+                    resolve(data);
+                }
+            );
+        });
+    }
+
+    /**
+     * プロフィール更新
+     */
+    public async updateProfile(params: {
+        accessToken: string;
+        profile: {
+            additionalProperty?: { name: string; value: string }[];
+            email?: string;
+            givenName?: string;
+            familyName?: string;
+            telephone?: string;
+        };
+    }): Promise<void> {
+        const { accessToken, profile } = params;
+        const userAttributes: AWS.CognitoIdentityServiceProvider.AttributeListType =
+            [];
+        Object.keys(profile).forEach((key) => {
+            if (key === 'additionalProperty') {
+                return;
+            }
+            if (key === 'email') {
+                userAttributes.push({ Name: 'email', Value: profile.email });
+            }
+            if (key === 'givenName') {
+                userAttributes.push({
+                    Name: 'given_name',
+                    Value: profile.givenName,
+                });
+            }
+            if (key === 'familyName') {
+                userAttributes.push({
+                    Name: 'family_name',
+                    Value: profile.familyName,
+                });
+            }
+            if (key === 'telephone') {
+                userAttributes.push({
+                    Name: 'phone_number',
+                    Value: profile.telephone,
+                });
+            }
+        });
+        if (profile.additionalProperty !== undefined) {
+            profile.additionalProperty.forEach((a) => {
+                if (a.name.match(/custom\:/) !== null) {
+                    userAttributes.push({
+                        Name: a.name,
+                        Value: a.value,
+                    });
+                }
+            });
+        }
+        await this.updateUserAttributes({ accessToken, userAttributes });
+    }
+
+    /**
+     * ユーザー削除
+     */
+    public async deleteUser(params: { accessToken: string }): Promise<void> {
+        return new Promise((resolve, reject) => {
+            AWS.config.update({ region: AwsCognitoService.REGION });
+            const { accessToken } = params;
+            const cognitoIdentityServiceProvider =
+                new AWS.CognitoIdentityServiceProvider();
+            cognitoIdentityServiceProvider.deleteUser(
+                {
+                    AccessToken: accessToken,
+                },
+                (err, _data) => {
+                    if (err && err.name !== 'NotAuthorizedException') {
+                        reject(err);
+                        return;
+                    }
+                    resolve();
+                }
+            );
+        });
     }
 
     /**
@@ -53,9 +234,11 @@ export class AwsCognitoService {
      * @returns {boolean}
      */
     public isAuthenticate(): boolean {
-        return (this.credentials !== undefined
-            && this.credentials.identityId !== undefined
-            && this.credentials.identityId.length > 0);
+        return (
+            this.credentials !== undefined &&
+            this.credentials.identityId !== undefined &&
+            this.credentials.identityId.length > 0
+        );
     }
 
     /**
@@ -73,17 +256,23 @@ export class AwsCognitoService {
         }
         await this.credentials.getPromise();
         const cognitoSync = new AWS.CognitoSync({
-            credentials: this.credentials
+            credentials: this.credentials,
         });
-        const listRecords = await cognitoSync.listRecords({
-            DatasetName: args.datasetName,
-            IdentityId: this.credentials.identityId,
-            IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID,
-            LastSyncCount: 0
-        }).promise();
-        if (listRecords.DatasetSyncCount === undefined
-            || listRecords.SyncSessionToken === undefined) {
-            throw new Error('listRecords: Records or DatasetSyncCount or SyncSessionToken is undefined');
+        const listRecords = await cognitoSync
+            .listRecords({
+                DatasetName: args.datasetName,
+                IdentityId: this.credentials.identityId,
+                IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID,
+                LastSyncCount: 0,
+            })
+            .promise();
+        if (
+            listRecords.DatasetSyncCount === undefined ||
+            listRecords.SyncSessionToken === undefined
+        ) {
+            throw new Error(
+                'listRecords: Records or DatasetSyncCount or SyncSessionToken is undefined'
+            );
         }
         if (listRecords.Records === undefined) {
             listRecords.Records = [];
@@ -92,13 +281,18 @@ export class AwsCognitoService {
         const mergeValue = this.convertToObjects(listRecords.Records);
         Object.assign(mergeValue, args.value);
 
-        const updateRecords = await cognitoSync.updateRecords({
-            DatasetName: args.datasetName,
-            IdentityId: this.credentials.identityId,
-            IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID,
-            SyncSessionToken: listRecords.SyncSessionToken,
-            RecordPatches: this.convertToRecords(mergeValue, listRecords.DatasetSyncCount)
-        }).promise();
+        const updateRecords = await cognitoSync
+            .updateRecords({
+                DatasetName: args.datasetName,
+                IdentityId: this.credentials.identityId,
+                IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID,
+                SyncSessionToken: listRecords.SyncSessionToken,
+                RecordPatches: this.convertToRecords(
+                    mergeValue,
+                    listRecords.DatasetSyncCount
+                ),
+            })
+            .promise();
         if (updateRecords.Records === undefined) {
             updateRecords.Records = [];
         }
@@ -111,26 +305,28 @@ export class AwsCognitoService {
      * @param {string} datasetName
      * @returns {Promise<any>}
      */
-    public async getRecords(args: { datasetName: string; }): Promise<any> {
+    public async getRecords(args: { datasetName: string }): Promise<any> {
         if (this.credentials === undefined) {
             throw new Error('credentials is undefined');
         }
         await this.credentials.getPromise();
         const cognitoSync = new AWS.CognitoSync({
-            credentials: this.credentials
+            credentials: this.credentials,
         });
-        const listRecords = await cognitoSync.listRecords({
-            DatasetName: args.datasetName,
-            IdentityId: this.credentials.identityId,
-            IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID,
-            LastSyncCount: 0
-        }).promise();
+        const listRecords = await cognitoSync
+            .listRecords({
+                DatasetName: args.datasetName,
+                IdentityId: this.credentials.identityId,
+                IdentityPoolId: AwsCognitoService.IDENTITY_POOL_ID,
+                LastSyncCount: 0,
+            })
+            .promise();
         if (listRecords.Records === undefined) {
             listRecords.Records = [];
         }
-        console.log('getRecords', this.convertToObjects(listRecords.Records));
+        // console.log('getRecords', this.convertToObjects(listRecords.Records));
 
-        return (<any>this.convertToObjects(listRecords.Records));
+        return <any>this.convertToObjects(listRecords.Records);
     }
 
     /**
@@ -139,7 +335,10 @@ export class AwsCognitoService {
      * @param {number} count
      * @returns {{ Key: string; Op: string; SyncCount: number; Value: string; }[]}
      */
-    private convertToRecords(value: any, count: number): {
+    private convertToRecords(
+        value: any,
+        count: number
+    ): {
         Key: string;
         Op: string;
         SyncCount: number;
@@ -150,7 +349,7 @@ export class AwsCognitoService {
                 Key: key,
                 Op: 'replace',
                 SyncCount: count,
-                Value: JSON.stringify(value[key])
+                Value: JSON.stringify(value[key]),
             };
         });
     }
@@ -163,16 +362,17 @@ export class AwsCognitoService {
      */
     private convertToObjects(records: any[]): Object {
         const result: any = {};
-        records.forEach((record: {
-            Key: string;
-            Op: string;
-            SyncCount: number;
-            Value: string;
-        }) => {
-            result[record.Key] = JSON.parse(record.Value);
-        });
+        records.forEach(
+            (record: {
+                Key: string;
+                Op: string;
+                SyncCount: number;
+                Value: string;
+            }) => {
+                result[record.Key] = JSON.parse(record.Value);
+            }
+        );
 
         return result;
     }
-
 }